requirements.txt
scapy>=2.4.5
paramiko>=2.7.2
python-nmap>=0.6.1
rich>=10.0.0
flask>=2.0.0
flask-restful>=0.3.9
blessed>=1.21.0
plotext>=5.3.2
requests>=2.25.1